requirements.txt
streamlit==1.14.0
pandas
plotly.express
wordcloud
spacy
sentiment-analysis-spanish
matplotlib.pyplot
keras 
tensorflow